wasm: don't access deleted settings objects
The emscripten_idb_async_ functions are async, so there
is no guarantee that the QSettings object which made
the request will be live by the time the completion
callback is made.

Keep track of live QWasmSettingsPrivate objects, return
early from the callbacks if userData does not point
to a valid QWasmSettingsPrivate.

Pick-to: 6.4 6.3 6.2 5.15
Change-Id: Ia319b1875dcf2c329ba27954e3f026e004fe0aac
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>
(cherry picked from commit 25c2d05340eee01cf55457b8327f8f69d408879a)
Conflicts:
	src/corelib/io/qsettings_wasm.cpp
(cherry-picked to 5.15.2+wasm as it fixes some crashes when trying to use the
QtFilePicker, even if that ultimately doesn't work anyway)