Minimal support for Emscripten -sPROXY_TO_PTHREAD=1
...with lots of TODOs still for how to approximate existing functionality in
that case, and there may well be more cases where the code expects that there's
a global JS window object or that the JS canvas object is a HTMLCanvasElement,
both of which are not true in this case.

Client code actually making use of -sPROXY_TO_PTHREAD=1 will want to set the new
Module.ignoreApplicationExit = true now, as in the PROXY_TO_PTHREAD case the
application exits very early by design (though its code still keeps on running),
and qtloader.js would then hide the canvas and show a meaningless "Application
exit with code 0" in its place.

(seen working with LO built with -sPROXY_TO_PTHREAD=1
-sOFFSCREENCANVAS_SUPPORT=1 -sOFFSCREENCANVASES_TO_PTHREAD=\#qtcanvas)