Revert "Frontend for headless pdf conversion in qtwasm"
This reverts commit b345732eaa6d9f57508a8d18523c75f51d2a7bb3.  It is some kind
of experiment that had been "parked" on this 5.15.2+wasm branch past the
v5.15.2+wasm tag.  But now we need LO to use further fixes (that have also
already been submitted to this branch past that tag), but not this experimental
commit, and the easiest thing is to keep using this branch, make LO use its
HEAD, and just revert this problematic experimental commit.