Avoid increasing canvas to a way-to-large size
For example with
<https://github.com/allotropia/zetajs/tree/main/examples/letter-address-tool>, I
experienced poor results where the canvas was enlarged to an excessive size and
scaling, showing only a top-left fraction of the Writer document, without even a
way to scroll or to show the list of addresses that was moved off the right edge
of the browser window.

On a Fedora 41 GNOME Wayland machine with a 2880x1800 screen and scaling set to
175%, I experienced this issue with firefox-132.0-2.fc41.x86_64, but not with
google-chrome-stable-130.0.6723.116-1.x86_64.  On a macOS 15.1 machine with a
3840x2160 screen set to a 1920x1080 resolution, I experienced this issue with
all of Chrome 130.0.6723.117, Firefox 132.0,1, and Safari 18.1.

In all the scenarios where I experienced the issue,
QWasmScreen::devicePixelRatio was returning 2.0, while in the scenario where
everything was fine, it was returning 1.0.  (In each case, it would return these
values for both the original code that does the dpr*scale computation, as well
as for the PROXY_TO_PTHREAD case calling emscripten_get_device_pixel_ratio.)
And always using 1.0 appears to produce the best results everywhere.  (Letting
QWasmScreen::devicePixelRatio still return 2.0, and instead removing the
`cssSize * devicePixelRatio()` scaling from
QWasmScreen::updateQScreenAndCanvasRenderSize would lead to a good canvas size,
but would still cause its content to be scaled up, so that only part of it would
be visible.)